Move books to the redux folder
