+++ src/Redux/books/BookList.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { v4 as uuidv4 } from 'uuid';
-import BookItem from '../Redux/books/BookItem';
+import BookItem from './BookItem';
 
 const BookList = () => {
   const booksDetails = [
